made the CX10Command values range from -1 to 1. Switched to junit

+++ lib/src/main/java/schuitj/drone/lib/drone/Drone.java
@@ -1,13 +1,13 @@
 package schuitj.drone.lib.drone;
 
 public interface Drone {
-    void setThrottle(int amount);
+    void setThrottle(float amount);
 
-    void setPitch(int amount);
+    void setPitch(float amount);
 
-    void setYaw(int amount);
+    void setYaw(float amount);
 
-    void setRoll(int amount);
+    void setRoll(float amount);
 
     void takeOff();
